admin/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==24.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
Pillow==10.4.0
